extract validations of player_choice for opponent options to the validator class
