+++ src/constants.py
@@ -2,6 +2,7 @@ EXIT = "exit"
 WELCOME = "welcome"
 CATS = "finished"
 ERROR = "error"
+OPTION_ERROR = "option error"
 WIN = "win"
 MOVE = "move"
 REQUEST_MOVE = "request"
