+++ src/three_by_three_validator.py
@@ -15,3 +15,13 @@ class ThreeByThreeValidator():
         valid_input = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "exit"]
         return move in valid_input  
 
+    def get_valid_player_choice(self, cli):
+        valid_choices = ["1", "2", "3", "exit"]
+        choice = None
+        while not choice in valid_choices:
+            if choice != None: self.report_error(cli)
+            choice = cli.get_opponent()
+        return choice
+
+    def report_error(self, cli):
+        cli.invalid_option()
