+++ test/player_builder_test.py
@@ -6,17 +6,17 @@ from unittest.mock import Mock
 from src.player_builder import PlayerBuilder
 from src.human_player import HumanPlayer
 from src.computer_player import ComputerPlayer
+from src.three_by_three_validator import ThreeByThreeValidator
 from src.cli import Cli
 
 class PlayerBuilderTest(unittest.TestCase):
 
     def test_build_players_builds_list_of_two_human_players(self):
-        validator = Mock()
+        validator = Mock(ThreeByThreeValidator)
         processor = Mock()
         cli = Mock()
         builder = PlayerBuilder(validator, processor, cli)
-        builder.get_valid_choice = Mock()
-        builder.get_valid_choice.return_value = "1"
+        validator.get_valid_player_choice.return_value = "1"
         tokens = ['X', 'O']
         expected = True
         expected_size = 2
@@ -29,12 +29,11 @@ class PlayerBuilderTest(unittest.TestCase):
         self.assertEqual(expected_size, actual_size)
 
     def test_build_players_builds_list_of_one_human_one_computer(self):
-        validator = Mock()
+        validator = Mock(ThreeByThreeValidator)
         processor = Mock()
         cli = Mock()
         builder = PlayerBuilder(validator, processor, cli)
-        builder.get_valid_choice = Mock()
-        builder.get_valid_choice.return_value = "2"
+        validator.get_valid_player_choice.return_value = "2"
         tokens = ['X', 'O']
         expected = True
         expected_size = 2
@@ -49,12 +48,11 @@ class PlayerBuilderTest(unittest.TestCase):
         self.assertEqual(expected_size, actual_size)
 
     def test_build_players_builds_list_of_human_and_computer(self):
-        validator = Mock()
+        validator = Mock(ThreeByThreeValidator)
         processor = Mock()
         cli = Mock()
         builder = PlayerBuilder(validator, processor, cli)
-        builder.get_valid_choice = Mock()
-        builder.get_valid_choice.return_value = "3"
+        validator.get_valid_player_choice.return_value = "3"
         tokens = ['X', 'O']
         expected = True
         expected_size = 2
@@ -81,44 +79,4 @@ class PlayerBuilderTest(unittest.TestCase):
         actual = builder.build_players(tokens)
 
         self.assertEqual(expected, actual)
-
-    def test_get_valid_choice_returns_valid_choice(self):
-        validator = Mock()
-        processor = Mock()
-        cli = Mock(Cli)
-        choice = "1"
-        cli.get_opponent.return_value = choice
-        builder = PlayerBuilder(validator, processor, cli)
-        expected = choice
-
-        actual = builder.get_valid_choice()
-
-        self.assertEqual(expected, actual)
-
-    def test_get_valid_choice_returns_exit_when_chosen(self):
-        validator = Mock()
-        processor = Mock()
-        cli = Mock(Cli)
-        choice = constants.EXIT
-        cli.get_opponent.return_value = choice
-        builder = PlayerBuilder(validator, processor, cli)
-        expected = choice
-
-        actual = builder.get_valid_choice()
-
-        self.assertEqual(expected, actual)
-
-    def test_get_valid_choice_only_returns_valid_choice(self):
-        validator = Mock()
-        processor = Mock()
-        cli = Mock(Cli)
-        choices = ["gibberish", constants.EXIT]
-        cli.get_opponent.side_effect = choices
-        builder = PlayerBuilder(validator, processor, cli)
-        expected = constants.EXIT
-
-        actual = builder.get_valid_choice()
-
-        self.assertEqual(expected, actual)
-
