+++ test/three_by_three_validator_test.py
@@ -1,15 +1,18 @@
 import unittest
+import src.constants as constants
 
-from src.three_by_three_validator import ThreeByThreeValidator as Validator
+from unittest.mock import Mock
+from src.three_by_three_validator import ThreeByThreeValidator
 from src.three_by_three_board import ThreeByThreeBoard
 from src.three_by_three_processor import ThreeByThreeProcessor
+from src.cli import Cli
 
 class ThreeByThreeValidatorTest(unittest.TestCase):
     def test_validate_returns_proper_action_when_move_is_valid(self):
         move = "1"
         board = ThreeByThreeBoard()
         processor = ThreeByThreeProcessor(board)
-        validator = Validator(processor)
+        validator = ThreeByThreeValidator(processor)
         expected = ["move", "1"]
 
         actual = validator.validate(move)
@@ -20,7 +23,7 @@ class ThreeByThreeValidatorTest(unittest.TestCase):
         board = ThreeByThreeBoard()
         move = "Gibberish"
         processor = ThreeByThreeProcessor(board)
-        validator = Validator(processor)
+        validator = ThreeByThreeValidator(processor)
         expected = ["error", "gibberish"]
 
         actual = validator.validate(move)
@@ -31,7 +34,7 @@ class ThreeByThreeValidatorTest(unittest.TestCase):
         board = ThreeByThreeBoard()
         move = "87"
         processor = ThreeByThreeProcessor(board)
-        validator = Validator(processor)
+        validator = ThreeByThreeValidator(processor)
         expected = ["error", "87"]
 
         actual = validator.validate(move)
@@ -42,7 +45,7 @@ class ThreeByThreeValidatorTest(unittest.TestCase):
         board = ThreeByThreeBoard()
         move = "gibberish"
         processor = ThreeByThreeProcessor(board)
-        validator = Validator(processor)
+        validator = ThreeByThreeValidator(processor)
         expected = False
 
         actual = validator.validate_input(move)
@@ -53,10 +56,46 @@ class ThreeByThreeValidatorTest(unittest.TestCase):
         board = ThreeByThreeBoard()
         move = "6"
         processor = ThreeByThreeProcessor(board)
-        validator = Validator(processor)
+        validator = ThreeByThreeValidator(processor)
         expected = True
 
         actual = validator.validate_input(move)
 
         self.assertEqual(expected, actual)
-        
+
+    def test_get_valid_player_choice_returns_valid_choice(self):
+        processor = Mock()
+        validator = ThreeByThreeValidator(processor)
+        cli = Mock(Cli)
+        choice = "1"
+        cli.get_opponent.return_value = choice
+        expected = choice
+
+        actual = validator.get_valid_player_choice(cli)
+
+        self.assertEqual(expected, actual)
+
+    def test_get_valid_choice_returns_exit_when_chosen(self):
+        processor = Mock()
+        validator = ThreeByThreeValidator(processor)
+        cli = Mock(Cli)
+        choice = constants.EXIT
+        cli.get_opponent.return_value = choice
+        expected = choice
+
+        actual = validator.get_valid_player_choice(cli)
+
+        self.assertEqual(expected, actual)
+
+    def test_get_valid_choice_only_returns_valid_choice(self):
+        processor = Mock()
+        validator = ThreeByThreeValidator(processor)
+        cli = Mock(Cli)
+        choices = ["gibberish", constants.EXIT]
+        cli.get_opponent.side_effect = choices
+        expected = constants.EXIT
+
+        actual = validator.get_valid_player_choice(cli)
+
+        self.assertEqual(expected, actual)
+
